ml/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
prophet==1.1.6
statsmodels==0.14.4
numpy==2.1.3
python-dotenv==1.0.1
